Delete works (mostly)
The "mostly" refers to the fact that delete does not currently return
any information in the response to indicate the delete was successful
aside from a 200 code. For now I think this is okay, but ideally I'd
like to return something in the body, perhaps what was in the deleted
file?